pyproject.toml
[tool.poetry]
name = "tradingresults"
version = "0.1.0"
description = ""
authors = ["anna putilovskaya <putilovskayaam@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
asyncio = "^3.4.3"
aiohttp = "^3.11.18"
bs4 = "^0.0.2"
pandas = {extras = ["excel"], version = "^2.2.3"}
fastapi = {extras = ["standard"], version = "^0.115.12"}
asyncpg = "^0.30.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.40"}
alembic = "^1.15.2"
python-dotenv = "^1.1.0"
pydantic-settings = "^2.9.1"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.2"}
apscheduler = "^3.11.0"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
flake8 = "^7.2.0"
pytest-asyncio = "^0.26.0"
pytest = "^8.3.5"
asgi-lifespan = "^2.1.0"
pytest-cov = "^6.1.1"
aioresponses = "^0.7.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
